+++ tests/recorder/test_response_formatter.py
@@ -1,8 +1,6 @@
 from baby_steps import given, then, when
 
-from vedro_httpx.har import SyncHARFormatter
-
-from .._utils import (
+from tests._utils import (
     HTTPClientType,
     RouterType,
     build_response,
@@ -12,6 +10,7 @@ from .._utils import (
     sync_httpx_client,
     sync_transport,
 )
+from vedro_httpx.recorder import SyncHARFormatter
 
 __all__ = ("sync_formatter", "sync_httpx_client", "sync_transport", "builder",
            "respx_mock",)  # fixtures
