+++ vedro_httpx/__init__.py
@@ -1,5 +1,4 @@
 from ._async_http_interface import AsyncClient, AsyncHTTPInterface
-from ._request_recorder import RequestRecorder
 from ._response import Response
 from ._sync_http_interface import SyncClient, SyncHTTPInterface
 from ._vedro_httpx import VedroHTTPX, VedroHTTPXPlugin
@@ -7,4 +6,4 @@ from ._version import version
 
 __version__ = version
 __all__ = ("VedroHTTPX", "VedroHTTPXPlugin", "SyncHTTPInterface", "AsyncHTTPInterface",
-           "SyncClient", "AsyncClient", "Response", "RequestRecorder")
+           "SyncClient", "AsyncClient", "Response",)
